requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.7.0
asttokens==3.0.0
cffi==1.17.1
cfgv==3.4.0
cli_helpers==2.3.1
click==8.1.8
configobj==5.0.9
cryptography==44.0.0
decorator==5.1.1
distlib==0.3.9
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
executing==2.1.0
fastapi==0.115.6
filelock==3.17.0
greenlet==3.1.1
h11==0.14.0
identify==2.6.6
idna==3.10
ipython==8.31.0
isort==5.13.2
jedi==0.19.2
Mako==1.3.8
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
nodeenv==1.9.1
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pgcli==4.1.0
pgspecial==2.1.3
platformdirs==4.3.6
pre_commit==4.1.0
prompt_toolkit==3.0.48
psycopg==3.2.3
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.18.0
PyJWT==2.8.0
python-decouple==3.8
python-jose==3.3.0
python-multipart==0.0.20
PyYAML==6.0.2
rsa==4.9
setproctitle==1.3.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlmodel==0.0.22
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.41.3
tabulate==0.9.0
traitlets==5.14.3
typing_extensions==4.12.2
uvicorn==0.34.0
virtualenv==20.29.1
wcwidth==0.2.13
